pyproject.toml
[project]
name = "pydantic-craft-meet"
version = "0.1.0"
description = "A Python project generated from the boilerplate template"
requires-python = ">=3.13"
authors = [
    { name = "Bernardo", email = "your.email@example.com" }
]

dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    "fastapi[standard]>=0.115.0",

    # For Gemini structured outputs (Unit 8)
    "google-genai>=1.0.0",
]


[project.optional-dependencies]
dev = [

    "httpx>=0.27.0",

    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
# CLI entry point - uncomment and modify as needed
# pydantic-craft-meet = "pydantic_craft_meet.hello:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["units"]


# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
line-length = 88
target-version = "py313"
src = ["units"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]

# Example ignores - customize based on your team's preferences
ignore = [
    # Docstring rules - uncomment to enforce docstrings
    # "D100",   # Missing docstring in public module
    # "D103",   # Missing docstring in public function
    # "D104",   # Missing docstring in public package

    # Line length - sometimes long lines are acceptable
    "E501",   # Line too long (handled by formatter)

    # These conflict with formatter
    "W191",   # Indentation contains tabs
    "E111",   # Indentation is not a multiple of four
    "E114",   # Indentation is not a multiple of four (comment)
    "E117",   # Over-indented

    # Allow non-abstract empty methods in abstract base classes
    "B027",

    # Too many arguments/statements - adjust thresholds in [tool.ruff.lint.pylint]
    "PLR0913",  # Too many arguments
    "PLR0915",  # Too many statements
]

[tool.ruff.lint.per-file-ignores]
# Allow assert in tests
"tests/**/*" = ["S101", "ARG", "PLR2004"]
# Allow assert in unit tests
"units/**/test_*.py" = ["S101", "ARG", "PLR2004", "PLC0415"]
# Allow common patterns in training exercises and solutions
"units/**/*.py" = ["PLR2004", "PLC0415", "PLR0912", "PLW2901", "UP046", "UP042", "PLW0603", "PLW0602", "B008", "E721", "F841"]
# Allow incomplete code, unused imports in exercises
"units/**/exercise.py" = ["ERA001", "ARG", "B018", "F401", "F841", "W605"]
# Allow unused imports in __init__.py
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]

known-first-party = ["units"]


[tool.ruff.lint.pylint]
max-args = 8
max-statements = 50

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# =============================================================================
# MyPy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "units.*"
# Training code demonstrates Pydantic features (type coercion, aliases, computed fields)
# that mypy doesn't understand well, so we disable strict checking
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
ignore_errors = true


# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["units"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# =============================================================================
# Coverage Configuration
# =============================================================================
[tool.coverage.run]
source = ["units"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true
